% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoefficients.R
\name{plotCoefficients}
\alias{plotCoefficients}
\title{plot coefficients of a gene given a DESeqDataSet (DESeq2 > v1.16)}
\usage{
plotCoefficients(dds, geneName, legend = TRUE)
}
\arguments{
\item{dds}{DESeq2 object (DESeqDataSet)}

\item{geneName}{name of the gene to be plotted, needs to be in the rownames of dds}

\item{legend}{(TRUE/FALSE, default TRUE), whether or not to plot the legend inside the plot.
the position is chosen automatically based on the last coefficient to be plotted.}
}
\value{
returns nothing
}
\description{
\code{plotCoeficients} plot coefficients of a gene given a DESeqDataSet (DESeq2 > v1.16)
}
\examples{
dds <- DESeq2::makeExampleDESeqDataSet()
dds = DESeq2::DESeq(dds)
plotCoefficients(dds, "gene930")

}
